store: do not defer LocalStoreCachedBackingStore caching
Summary:
In D48925204, it was observed that in some cases a very large amount of ready
SemiFuture could be queued but not executed due to not being attached to an
executor. This led to EdenFS memory usage growing almost unbounded which
resulted in OOM kills.

While the original diff fixed that issue by recognizing that de-duplication
could be done more effectively, the LocalStoreCachedBackingStore can be
improved slightly to avoid queueing when SemiFuture are ready. Fixing this is
trivial: merely converting these to ImmediateFuture will extract the value from
the SemiFuture and continue the computation immediately, thus avoiding
queueing.

Reviewed By: jdelliot

Differential Revision: D48850788

fbshipit-source-id: 1e56a3d1ef40b6a9e3cc69feeb12be13e02ddc69