proxy: quote url on Windows
Summary:
Previously, the code runs:

  cmd.exe /c start https://example.com/?a=b&c=d

which will open the `https://example.com/?a=b` URL without the `c=d` parameter,
since `&` is seen as a command separator. This breaks `cwd=`.

This diff changes the command to:

  cmd.exe /c "start "ISL" "https://example.com/?a=b&c=d""

to include the `c=b` parameter properly. The `"ISL"` is necessary since `start`
treats the first `"` as "title".

Further more, we can simplify the `cmd.exe /c` part with `shell: true` from
nodejs.

Reviewed By: evangrayk

Differential Revision: D50094506

fbshipit-source-id: 7976a97c920af20cfd2019dcfb50f05980b629e3