debugcurrentexe: command to show the path to the main executable
Summary:
This command prints the path to the main executable. It's similar to
`sys.executable` in `debugshell`, but without using Python.

The main motivation is to better support the node IPC channel on Windows.
The problem is that by default, libc (CRT, C runtime) fds (not confused with
Windows OS handles) are *not* inherited. To inherit them you'll need to set
the `lpReserved2` in `STARTUPINFOA` struct when creating a process. The field
is [documented](https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/ns-processthreadsapi-startupinfoa) as:

  cbReserved2
  Reserved for use by the C Run-time; must be zero.

  lpReserved2
  Reserved for use by the C Run-time; must be NULL.

Rust stdlib does not provide a way to set them in the C runtime compatible way
to pass C fds. nodejs/libuv sets them directly, ignoring the "must be NULL"
suggestions.

On Windows, there are multiple layers of wrappers today.

For the telemetry wrapper we might teach it to translate the fd to Windows HANDLE
and teach the main binary to also accept a HANDLE.

Reviewed By: sggutier

Differential Revision: D50020026

fbshipit-source-id: 1798a30848b2b76add63f5637594784e6e7e096c