Add "compact" mode option
Summary:
Add an option to render commits in compact mode, which mostly just makes commits with diffs attached not wrap into two lines.

This helps address the feeling that the new ISL is less compact at showing the same information as the old ISL. This is true, the new ISL tries to render diffs a bit "nicer", but at the expense of space usage. New users may not care and the better wrapping is probably a better default. But users with very large stacks may prefer to intentionally ensure the information is more compact.

We may use this setting for further space saving measures in the future. For now, I'm mostly just changing the wrapping and one small bit of reduced gap between commits.

Reviewed By: sggutier

Differential Revision: D50100194

fbshipit-source-id: cf53a7b0dec76aca72c7ce63b2ae579b373c1953