Support bounded export paths file arg
Summary:
## Context
See summary of GitExportGraphInfo.

## This diff
Exposes a way for users to specify export paths associated with HEAD commits (i.e. the latest commit where that path should be exported).

The CLI will still take a changeset that will be the default HEAD for all changesets provided by the `-p` flag and at least one path through the `-p` flag.

But now there's an optional argument where the user can provide a JSON file containing paths to export and their respective heads.

An example would be:
```
[
  {
    "path": "fbcode/scm/mononoke",
    "upper_bound_changeset": {
      "ID": "e1796eaaa74b94bde985a3b8ea4e3aff9045211c"
    }
  }
]

```

This can be used if a user sees a warning about a possible rename commit.

## TODO:
- Update integration test to confirm that a warning is printed when there's a rename and it isn't followed.

Reviewed By: mitrandir77

Differential Revision: D49685528

fbshipit-source-id: 2e9d4484534369adca0c057ef843d89e5cba7572