NewAdmin command to generate git bundle from Mononoke repo
Summary: This diff adds the functionality of generating a Git bundle for a given Mononoke repo based on the logic and data structures that have been implemented until now. The logic for generating the bundle lies completely in the NewAdmin command for now but would later be moved into a separate library of its own where it can be used by other components (e.g. `GitProtocol`, `GitExport` etc.). This diff does not contain any test, that will be added in follow-up diffs.

Reviewed By: mitrandir77

Differential Revision: D49636552

fbshipit-source-id: 3f2928049d817e72b4f26461fa58a89e72e71a30