Decrease padding on the outside of the comparison view modal
Summary: The border around the comparison view modal helps give some context to the modal. But at 40px on each side, we lose 80px of height with which we could show lines of code. Let's halve the padding around the modal, at least for the comparison view if not other modals.

Reviewed By: muirdm

Differential Revision: D50102348

fbshipit-source-id: e8d7b5eb7281a83053422ace3ebb3aeb610e4905