Make comparison context not generic
Summary: Minor refactoring. The Context we pass to the comparison view SplitDiffView is a bit weird since it was made for compatibility with reviewstack. It has an ID field, which is used in some atoms to key some cached data. This doesn't really need to be generic for our use case, since we always have a path and comparison. This simplifies the code a good amount.

Reviewed By: muirdm

Differential Revision: D50103450

fbshipit-source-id: dd3a90582b592651da5589a5ac43e0212fc263d1