Fix file header name overflowing
Summary:
Make the file name in the comparison view (& split view) scroll x so that you can always scroll to read long names.

Maybe it would be a little better to overflow on the left with ellipsis like we do for other file paths. But for now, I think just scrolling is OK.

Reviewed By: muirdm

Differential Revision: D50104126

fbshipit-source-id: 74bad16eec895ac9c365843800f3c2f6f423808a