status: fix case folding for watchman
Summary:
Revamp the Rust watchman status code path to handle case collisions on case insensitive filesystems.

We now do a pre-processing step on the watchman paths instead of post-processing. Basically we collapse "FOO" and "foo" into a single entry using the case of whatever is in treestate. The one exception is if "foo" is marked removed - in this case we keep them separate so we can manifest the status:

```
$ sl status
R foo
? FOO
```

Reviewed By: sggutier

Differential Revision: D50154149

fbshipit-source-id: 0c0c190696a7ccd60ab135e3d4c51be96c3bc056