Stop printing commit id from graphs used in integration tests
Summary:
The integration tests will save the graph information before and after the export (i.e. to compare the original repo with the temp repo). The files are then compared using `diff`.

Currently, the output is a bit confusing, specially when we're reviewing changes to the diff output, because any difference between the commits will lead to a different id and this really clutters the output.

Any relevant changes between the commits will already be showcased in the file changes list, so let's stop printing the commit ids to make the test outputs easier to understand.

FYI this won't be used for long, because I'll soon be using Rajiv's library to export the temp repo to a git repo, so I'll update the tests to inspect the git repo instead.

Reviewed By: mitrandir77

Differential Revision: D50171993

fbshipit-source-id: a5301fa60488227db5cd27f9451e98b52126324b