Separate chromelike_app and macos_app into different files
Summary:
Splitting the implementation for the macos webview app and the chromelike app makes it a little easier to dig through

I'm not so familiar with rust—is there a nice way we could just declare everything in this file to be macos only? I couldn't get the `#[cfg(target_os = "macos")]` to work for the entire file / mod.

Reviewed By: quark-zju

Differential Revision: D49841654

fbshipit-source-id: 3b217067da48e4be985637c8be5058470fe24423