Allow --foreground to open webview in the cli process
Summary:
Typically, we expect `sl web` to open the webview in its own mac application.

When developing, it can be useful to run the webview in the current process, so you don't have to deal with the separate app spawning indirection.

When doing this, it's nice to also show the log file contents in the stdout. This makes the webview behave like it's really in the foreground, and make it easier to debug server stuff.

Reviewed By: quark-zju

Differential Revision: D49938119

fbshipit-source-id: 1865ed8991c84375582b36c1b3cd9ac7c055a3ec