Add chromelike_app as a separate platform
Summary:
On windows, `sl web` opens chrome with --app. This should behave exactly like the browser platform.

If we use a separate platform name, we have flexibility to override some features for this use case in the future, plus our analytics&bug reporting will report this platform differently which will be useful.

Reviewed By: quark-zju

Differential Revision: D50091744

fbshipit-source-id: c2fa9211ee9a9ff340877c737c84f05ed0ad33a6