Support keyboard shortcuts in webview
Summary:
Support copy/paste/cut/select all/undo/redo in the webview. This only affects macOS right now, but later if we use webview for other platforms, this same issue may or may not be present.
According to https://github.com/webview/webview/issues/403, this is a macos thing. (note the differences are likely because of different built in OS webviews)

We can support these shortcuts manually. The outlier is pasting, which for some reason opens an OS popup to select "paste". This must be some security thing. Browsers are very particularly about only letting you do clipboard stuff in a secure context (events triggered by the user). Our code should fall in that category, but I guess the webview framework doesn't handle it quite correctly.

Reviewed By: quark-zju

Differential Revision: D50029233

fbshipit-source-id: 130c2e785e5277655c3fee26c1e5c2866c83389d