Add typed message passing between ISL and webview host
Summary:
From the rust code that spawns the webview, we need to be able to communicate with ISL for a few small bits of functionality. Currently this includes:
- opening external URLs
- showing alerts
- showing file picker dialogs

To do this, we pass JSON messages back and forth, with ID fields so the client can send requests that get responses.

I added support for two types of messages:
- "invoke" which sends a message with no answer
- "request" which sends a message and expects a single response

In this diff the "request" is just an example message which adds one to an argument. This will actually end up being an alert or confirm or file picker in later diffs.

Reviewed By: quark-zju

Differential Revision: D49837910

fbshipit-source-id: 2bc823cf295a8033f4c4c836f34dbaaeec57d54e