Make webview macOS app bundle spawn its own ISL server
Summary:
This is the start of a number of changes to actually use the new webview stuff for `sl web`.

This change lays the groundwork by refactoring the basic control flow.

Before webviews, `sl web` would spawn a node process, then that node process would open the browser

Before this diff, the way the new webview infra was set up, we woudl expect to have `sl web` in python spawn the isl server, then create a macOS app bundle pointing back to sl, then `open` that app bundle, while passing in cli args with the node server URL.

However, this architecture means that the app bundle is completely dependent on being called with the server args. If you launched sl web once, then saved the app icon in your dock, then closed and reopened, it would fail because it's not launched with the right URL in the process args.

The new design in this diff is to move the responsibility of spawning the node ISL server into the app bundle process.

So now `sl web` in python will determine the appropriate node args to spawn the isl server, then call into rust to construct the app bundle, then write all the node server args the `sl web` call got into a server_args.json file inside the app bundle, then call open on the app bundle.

Then, the newly spawned app will look for that server_args.json file, read it, and use those options to spawn a node ISL server.

This means it supports being reopened from the dock because it knows the last server spawn options so it can spawn its own server when it needs to.

There's still quite a lot more to do for this, but this is the minimal example that can open the webview from `sl web` invocation.

Some notable TODOs:
- revive the chrome --app code path which I sort of killed during this change
- add --browser to the server args
- add width&height to the server args
- check for --no-open and just spawn a regular server without opening the app
- actually save multiple windows worth of server spawn options to server_args.json, so we could support opening multiple windows in the same app
- Do some kind of gating so we can land this and control the rollout

Some more minor or later TODOs:
- eventually, move all the rest of the code out of isl.py and into rust (including the tar stuff)
- Fix some bugs when launched via webview, like syntax highlighting not working, no file dialogs, and no file uploads
- think about how to handle if the app is already open and you run `sl web` again
- do some version checking in the app so that if you leave the app open for a long time, it warns you to restart to update

P.S. my rust is a bit weak, let me know if I'm doing things in a silly way.

Reviewed By: quark-zju

Differential Revision: D49703009

fbshipit-source-id: df3405512f32138295946171b4a4b9bec43b55f3