Add escape hatch config to force --no-app
Summary:
Add a config bool that lets us make `--no-app` be the default for sl web. This way, we can roll out the webview / chromelike apps as a default and be confident that we can roll back if necessary.

I opted to do this in python since we have the config easily available. Eventually I think we should replace isl.py with a rust command and thus we'd need to do this in rust at that point.

Reviewed By: quark-zju

Differential Revision: D50043342

fbshipit-source-id: 11886ec50b53b95ce30891f9f69062bb7033f13b