Update blake2 to 0.10
Summary:
This updates blake2 to 0.10, as well as migrates all existing users.

This is a non-trivial diff and additional verification is appreciated.

Note that in 0.10, blake2 types were split into the MAC and non MAC variants. Migration for non-keying hash usage, such as in `zippydb`, has minimal migration steps.

For keyed hash usage (AKA using blake2's MAC functionality), migration is a little less clear. Since hashes can be used to lookup stable items, a change in hash output may lead to lost items. D49973500 was introduced to add snapshot tests to ensure that the hash value is consistent after an update.

The most notable change is that mononoke_types seems to run into a footgun with new_keyed -- blake2 0.9 does not pad an empty key to the correct size, while blake2 0.10 does. This due to the fact that in blake2 0.9, an empty key was interpreted to use the hash functionality of blake2 while a non-empty key was interpreted to use the MAC properties. This requires a change in tests.

Reviewed By: jsgf

Differential Revision: D49649655

fbshipit-source-id: 9a13e005ac998d068df8154de627e31b7fdf615e