Skip some data in UI state
Summary:
The UI state we log in bug reports includes everything from recoil, which includes selectors. This means we duplicate a lot of data.

For example, we have the entire list of commits, but also a derived Map of hash to subtree.

Let's remove any keys that produce noise or purely derived data that doesn't give any additional insights.

Note that I keep some stuff like treesWithPreviews AND commitTrees, since the difference between these is nontrivial.

Reviewed By: zzl0

Differential Revision: D50138497

fbshipit-source-id: 87b639ed0ac7586132b456e67211815822840e2e