Disable comparison view button on an optimistic commit
Summary:
Optimistic commits can't be opened in the comparison view currently. Maybe we could support that one day, but it's non trivial. Let's at least not show "OPTIMISTIC_COMMIT_HASH" in the comparison view button and disable the button

I played with adding a tooltip, but the time period for this is usually small, and the tooltip tends to get stuck if the source changes underneath it.

Reviewed By: cs01

Differential Revision: D50204084

fbshipit-source-id: 0d1d78b467728b56a5dfc8aa41a63fe3c4b17e5e