Make uncommit button less conspicuous
Summary:
The uncommit button was way to big for its purpose. It made it seem like we REALLY wanted you to uncommit. But it's not even a particularly common flow.

Here I made it an icon button instead, which is more subtle. The icon here is the opposite of the amend icon, so hopefully it seems clear that it's kind of like undoing commit. It points "up" towards the uncommitted changes list, which feels right to me.

Reviewed By: quark-zju

Differential Revision: D50200911

fbshipit-source-id: 70287cb28dc1c3de47c7e9ae159975a8c7f3338e