Add subtle background to icon buttons
Summary:
We've gotten misc feedback from folks that buttons are harder to see in ISL.

We use a lot of "icon" buttons, which can look like just text in some cases.

Adding a heavy background color can make them a bit overwhelming. So here I played with adding a very very subtle background color. It gives just a bit of separation with the background.

I've played with this idea in the past but never went ahead with it. I think it's probably worth doing something to make these icons easier to see.

I'm not sure I'm 100% happy with this. I do think it makes it a bit more clear what's a button and what's text.

Would love any feedback about this. Too subtle? not worth doing?

Reviewed By: quark-zju

Differential Revision: D50202130

fbshipit-source-id: 7c996c9e7ca596fd70db1bdeadcf56f9c1ff292d