Back out "store: deduplicate BackingStore operations in ObjectStore"
Summary:
Using `Promise::setTry` properly is tricky, in particular it's behavior depends
on the executor attached to the underlying `SemiFuture`/`Future`. In the case
where the executor is an inline executor (such as `folly::InlineExecutor` or
`folly::QueuedImmediateExecutor`), the continuation attached to the `Future`
will be executed immediately. In most cases this isn't an issue, but becomes
one when the continuation tries to acquire a lock that is already taken.  Here,
the sequence of events goes like:
  frame#14: ObjectStore::getTree at L210 <- tries to get lock that is already taken
  frame#16: getOrFindChildHelper at L610
  frame#17: VirtualInode::getOrFindChild at L642
  frame#24: VirtualInode::getOrFindChild at L635
  frame#25: VirtualInodeLoader::loaded at L96
  frame#28: VirtualInodeLoader::loaded at L95
  frame#29: VirtualInodeLoader::loaded at L104 <- recursive call to itself
  frame#54: QueuedImmediateExecutor::add at L33 <- execute callback inline
  frame#64: tryDedupGet at L223 <- SharedPromise::setTry
That is, `VirtualInodeLoader` tries to call `ObjectStore::getTree` in the
continuation leading to the deadlock.

Fixing this one is trivial (D50199539) by just moving the `setTry outside of the
lock, but it's unclear at this point if there are other code path that make use
of `QueuedImmediateExecutor` in a problematic manner. Backing out feels safer
at this point.

Reviewed By: mshroyer

Differential Revision: D50235920

fbshipit-source-id: 4744f177d8bae142f81387adf30f487bf9e3ef09