mercurial_types: skip re-uploading filenode pointers we know exist
Summary:
When deriving mercurial changesets, we use the filenode lookup cache to find filenode ids that already exist for known content and parents.  However, if we get a cache hit, we still re-upload the filenode lookup pointer, even though we know it was there.

Speed up derivation where filenode pointers are cache hits by skipping the re-upload.  Ideally we'd skip uploading the filenode entirely, as it should already exist, however we don't actually know that for sure, as something could've successfully uploaded the filenode pointer, but failed to upload the filenode itself.

Reviewed By: quark-zju

Differential Revision: D50014634

fbshipit-source-id: 818471e7a1ddd30b8420e3462d856de9bc690cdc