Remove pushrebase_disable_rebased_commit_validation tunable
Summary: The pushrebase_disable_rebased_commit_validation tunable was added to guard a new set of validations that we put in place to address problems in pushrebasing diamond merge commits. However, its been more than 3 years since that logic was enabled in production and we never disabled it once. Makes sense to get rid of this check completely

Reviewed By: mzr

Differential Revision: D50261925

fbshipit-source-id: c2bdec2d27a83821b294a27d3a4249de76b3296f