Remove max_scuba_msg_length tunable
Summary: We never changed this tunable since introducing it. I think it is reasonable to thik we will change it in the future because f how much space the message column takes up, but we can tweak it via an update of this constant. I don't think this needs to be in a config either

Reviewed By: mzr

Differential Revision: D50261199

fbshipit-source-id: dc7e0594a3499b0a3f090b2a5b3f6addece0c606