Remove bookmark_subscription_protect_master tunable & bookmark_subscription_max_age_ms tunable
Summary:
These were introduced in D28569840 for bookmark_subscription_protect_master to derisk the rollout and never changed since May 2021 and D28505644 for the same reason.

I have had to change a test since although in prod we would fail to delete master, our test used the bool default (false) allowing the test to succeed. Now it checks that master will NOT be deleted.

There are other test confirming deletion of bookmarks.

Reviewed By: mzr

Differential Revision: D50261599

fbshipit-source-id: 6567b5ac58c879e90b8dd611dd07b07a30c12e19