Add integration test for retryable update
Summary:
When enabled, the configuration added in D47099273 causes error conflicts
reported by EdenFS during update to leave the repository in an "unfinished
update" state.  The original intent of this configuration was to deal with
network errors, but this behavior also applies to errors caused by EdenFS being
unable to remove a locked placeholder file.

This adds integration tests to ensure that a successfully retried update can
bring the repository to the desired end state.  These tests reveal that that
resuming update to deal with a failed placeholder invalidation doesn't yet
work, so some assertions are currently commented out with a TODO.

Reviewed By: genevievehelsel

Differential Revision: D50250253

fbshipit-source-id: b88e47d91819167f5f079c676b4026f3e02f7606