Remove tunables controlling logging to edenapi_request_dump scuba table
Summary:
This diff removes two tunables controlling sampling rate for requests and threshold from which requests become unsampled.

Although we have changed this in the last 6 months, I don't expect us to need to change it frequently, and waiting for a Mononoke release to update these values is reasonable as it is related to our behaviour with logging and is not user facing.

Reviewed By: markbt

Differential Revision: D50266523

fbshipit-source-id: 71ad6cd9af191368a5015dc27ddc05f2b21ccd06