Get rid of filenodes disabled tunable
Summary:
We added a tunable to control whether filenodes would be read from mononoke_production XDB in the pull or read path to avoid SEVs like S200532 which brought Mononoke down because mononoke_production XDB was down. We never used it in the past 3 years and have since started relying more on more on mononoke_production database (e.g. commit-graph, bookmarks, etc.) to the point where it wouldn't really be possible to meaningfully serve traffic if XDB does go down. Thus, it makes no sense to have this tunable as well.

Initially I decided to also get rid of the `FilenodeResult` type but that was a can of worms I am willing to open.

Reviewed By: clara-9

Differential Revision: D50262809

fbshipit-source-id: a31098dbf10db26412de5139c7b41ba6771f1bf4