Replace filenodes_master_fallback_ratio with a corresponding JustKnob
Summary: We introduced filenodes_master_fallback_ratio tunable as a follow up to Sev S211322 where a bug in HG client led to manifest and filenode requests for incorrect paths. The replica DBs returned no values for these paths and since there was a fallback to master in such case, we started hammering the master DB with loads of requests affecting overall Mononoke performance. This diff replaces that tunable with a JK that we can use in such future SEV-like scenarios. It also simplies it so that we either enable it or disable it (instead of leaving it up to chance)

Reviewed By: markbt

Differential Revision: D50267995

fbshipit-source-id: f28976454ff4a6a62b845dacc5cdeeaab24cc789