inodes: remove use of QueuedImmediateExecutor in VirtualInodeLoader
Summary:
Using QueuedImmediateExecutor can lead to deadlocks as seen in D50199539, let's
instead make use of an actual executor that the caller passes. When called in
the context of a Thrift call, the Thrift executor is passed in as the rest of
the Thrift handling is already done in that executor.

Reviewed By: genevievehelsel

Differential Revision: D50227150

fbshipit-source-id: 042617afa1e05fe602807caab7b6419e18ba6848