Add context menu option to open containing folder
Summary:
Standalone/browser users especially might want to explore files in their OS file explorer. Add a context menu to open the containing folder for a file.

We don't do this in vscode since we have the file tree.

This doesn't play very well with a remote connection, since you just can't use your OS file explorer in that case. So if you tried it would just fail... but we don't have a real remoting system right now, just manual tunneling. So we have no way to detect and disable this feature.

Reviewed By: quark-zju

Differential Revision: D50274726

fbshipit-source-id: e0f2ab22c72c77c71ad7392c1af0ee076fd60207