Skeleton for library that generates stream of packfile input
Summary: Currently the entire code for generating bundles and packs resides directly in the `create-bundle` command in NewAdmin. However, we would want to use this functionality for multiple purposes and hence it needs to be abstracted out into a library. Since `git-protocol` is going to be its primary user and decide on its eventual shape, I have named this library/crate `protocol`.

Reviewed By: mitrandir77

Differential Revision: D50079964

fbshipit-source-id: 4ce3d548decf5179f8fd96f3a9a9146117eb2a05