Fixing issues with add/remove file workflow
Summary:
Mateusz's comment on the `create-bundle` command diff made me realize that we were not accounting for duplicate objects and preventing them from getting injected into the packfile/bundle. This test case has been created to test that scenario and ensure that the pack/bundle that gets generated in such cases is considered valid by the Git client. In particular, this tests the following cases:
- Add/Remove/Add files/tree does not result in the same object getting included more than once in the pack
- Add/Remove/Add files/trees does not result in double counting the objects in the pack header
- Add/Remove/Add files/trees does not result in chains of deltas

Reviewed By: markbt

Differential Revision: D50119954

fbshipit-source-id: 91c3ea3f54c676ca77d35d03c7097d307fbcea65