Core library implementation for generating pack input stream
Summary: This diff adds the core logic for generating the list of packfile items along with information about included refs and the total number of object included in the packfile. Most of it is borrowed from the `create_bundle` NewAdmin command but modified to make it more generic and reusable across multiple use cases.

Reviewed By: markbt

Differential Revision: D50091938

fbshipit-source-id: d65ab8f1bcd2e2898cbd0f5dd887967e3c56f0b7