Use HashMap and HashSet with faster hashing algorithm for pack stream generation
Summary: We use `HashSet` and `HashMap` while generating pack item stream to de-duplicate objects or refs. These collections can grow quite large in which case the performance of the hashing algorithm plays a role in the overall runtime. Since the input to these collections are known to be safe, we do not need a cryptographic hash and can instead use a much faster non-cryptographic hashing algorithm used within the rust compiler

Reviewed By: markbt

Differential Revision: D50127591

fbshipit-source-id: f3d819d621e690cf2480ba80a5b02894a2348de9