Support file change filter in commit rewrite API
Summary:
IIUC, getting the implicit deletes from a commit is O(file changes). We're currently doing that for all the file changes from a bonsai for every commit rewrite.

However, in the case of gitexport, many of the file changes are irrelevant, so they just take a lot of time unnecessarily.

This is a contributing factor to very large diffs (e.g. codemods) taking forever to be rewritten.

This diff introduces a way to filter file changes before the implicit deletes are fetched and the multi mover is called, to avoid processing changes to irrelevant paths.

It significantly improved the speed of gitexport (see test plan).

Not sure if this filter could also be used on other use cases of the `rewrite_changeset` function, but if it can, it would likely also be a big performance improvement for those cases...

Reviewed By: RajivTS

Differential Revision: D50228048

fbshipit-source-id: 2247a16967873ec419f9193c25c3ef926b25a59f