Simplify fetch for backup statuses
Summary:
quark-zju recommended a simpler fetch for commit cloud status, by using a cheaper revset than `smartlog() and draft()`: `draft() - backedup()`. This way we  don't bother sending successfully backed up commits (which is basically the default).

This means we can use a simpler template too, so we don't use a complicated "if" thing in the template, and instead just do it ourselves in typescript.

Reviewed By: quark-zju

Differential Revision: D50392960

fbshipit-source-id: 333b2cc8e8232d4ee864cbfe8f33f952e6867c95