Change current commit cloud workspace
Summary:
When you change the dropdown with the current commit cloud workspace, run `sl cloud switch -w ` to switch to the new workspace.

While that happens, we disable the dropdown. We also optimistically override the current workspace name to make it seem like the operation succeeded immediately. We'll need to refresh after the command.

Reviewed By: quark-zju

Differential Revision: D50333279

fbshipit-source-id: bac6576b5e4cf5321c6faaead952426ddc188179