Make runOperation return a Promise
Summary:
While queuing is powerful for running multiple things, sometimes we need to refresh data when an operation finishes. Thus it's helpful to be able to know when an operation completes.

We need to save a callback for the Operation so we can resolve it. We can't really store this in the operation itself, since the Operation is sometimes stored in atoms which are immutable (for example, `operationBeingPreviewed`). So we store a global mapping. This is similar to the operation list we keep to know which operation the server is talking about.

The use case in question here is to refresh the Commit Cloud state when the `sl cloud switch` command completes, so that we can for sure get the latest data for the new workspace you're in.

Without being able to wait for the operation, this would be difficult.

In the future, we could extend this Promise to also include information about the completed operation (perhaps just `OperationInfo`?)

Reviewed By: quark-zju

Differential Revision: D50333278

fbshipit-source-id: 7feec709e844f2cdc62b26b8cfac879954bfd843