Fix filenames with #
Summary:
We have some weird file path handling using rtl and unicode rtl marks so that files truncate on the left side. We had one preceding mark, but it looks like we also need a trailing mark.

Before, it handled filenames like `.gitignore` so they aren't rendered as `gitignore.`. But it wasn't working for a filename like `#backupfile#`, it would put the trailing `#` to the front.

Reviewed By: quark-zju

Differential Revision: D50421680

fbshipit-source-id: 3640f0f40e22e7650d2032018a1f66d0859eee89