extend usage of ConfigSetting::getIdx()
Summary:
`ConfigSetting::getIdx()` was added in D48758164.  I missed a few call sites where we cast a `ConfigSettingType` to a `size_t`, which breaks the configs when we reorder the priority.

This diff also moves `getIdx` to `ConfigSettingBase`, since calling it from `ConfigSetting` requires a template for the config value type, which is not needed for this function

Reviewed By: xavierd

Differential Revision: D50372252

fbshipit-source-id: 31afd76ce7fd4dee801bf9989976bb18690c8821