Add integration test for implicitly deletes
Summary:
FYI, this is the behaviour of the `rewrite_commit` function ([here](https://fburl.com/code/ap5o4oce)) and is not specific to gitexport.
I added the test to make sure I won't break this behaviour in further changes.

I basically test 2 different scenarios of implicit delete:

## 1. File added is also exported
The `bar` file implicitly deletes `bar/f/g` and `bar/h/i`. But because `bar` is also exported, the new commit will also have the implicit delete information, so there's no need to "augment" the new bonsai with DELETION information.
That's why the `FileChange::Deletion`s for this scenario are added [here](https://fburl.com/code/m9uwnud6) and then removed [here](https://fburl.com/code/1cv98w0j) by `minimize_file_change_set`.

## 2. File added is NOT exported
The `foo` file implicitly deletes `foo/a/c` and `foo/a/d`. However, `foo` is **not** exported.

This means that the 2 files that **are exported** will be deleted, but the addition of `foo` will not be in the new bonsai, meaning there will be nothing to convey the deletion of the files.

That's why we need to "augment" the new bonsai with the deletions, so in the new version of commit B you'll see DELETION file changes for both of the implicitly deleted files.

Reviewed By: markbt

Differential Revision: D50320259

fbshipit-source-id: 0ec41e13d243388c200d0ba65f105378eadffc3e