Improve performance of implicit_deletes module
Summary:
## Context
Implicit deletes for a given changeset are found by:
- Iterating over all the file changes and keeping only the ones that modify/add,
- For each of these file changes, lookup an entry in the parent's manifest for that path.
  - If an entry is found and it's a leaf, no implicit deletes.
  - If an entry is found and it's a tree, return all the leaf entries from that tree, i.e. all the files/directories under the directory being implicitly deleted.

## Problem
`find_entries` ([code](https://fburl.com/code/ugxvq3jv)) seems to more efficient way to do manifest lookups for multiple paths.

The current implementation iterates over each file change and looks up its entry using `find_entry`, which just calls `find_entries` passing a single path.

**Improvement #1:** Pass all the paths directly to `find_entries`.

The other problem is that the calls to `list_leaf_entries` returns a stream and the streams were being flattened using `try_flatten_stream`, which IIUC preserves the order of the streams and **executes them serially**.

**Improvement #2:** Use `try_flatten_unordered` to introduce concurrency in the execution of the streams.

#thanks mitrandir77 for helping me flatten streams concurrently and without forcing execution!

## Notes
I initially used 100 as the concurrency limit, but increasing it to 10000 led to a big improvement, so I decided to go with that for now.

`get_implicit_deletes_single_parent` gets called for each parent of the changeset. I'm assuming that most changesets don't have a high number of parents and, because there's already a high degree of concurrency in `get_implicit_deletes_single_parent`, I decided to set the concurrency limit to 10 when flattening the results of each parent.

Reviewed By: markbt

Differential Revision: D50083311

fbshipit-source-id: 04f20f691597001b0f60651a0b635bff2d6e1596