Ensure all trees are uploaded recursively when importing commits
Summary: When we added the code for importing raw Git trees while transforming Git commits to Mononoke commits through `gitimport`, we (and by "we" that I mean "me" :P) forgot to consider the recursive nature of trees which would have required us to import all subtrees that had changed in the current commit. This diff fixes that behavior. I will re-run gitimport for whatsapp repo to update our data stores.

Reviewed By: markbt

Differential Revision: D50409874

fbshipit-source-id: 1fe8fa14023cde211f24c64b50dc4e065e569756